# Remove checkBackloggedCursors from subscribe